poz2fgdb fix bug in relationship builder
Turns out, I was a little lazy in my code.  I had fully specified
tables in my relation definition, but the output relationship was not
fully specified, and while it must create the relationship in the same
database as the tables, it is not smart enough to default to the same
database.  In your case it was trying to create the relationship in
your default fgdb.  In this fix, I am fully specifying the path to the
relationships.