Make sync work over https on server
Cert and private key are not shared in the repo.  They are private and
specific to the server, so look there.  A fully qualified domain name
is required, and the cert is signed by DOI, so the client must have the
DOIRootCA2 installed.  This will be true for all iPads on the NPS
network.